libft/lambdas/i2c_push_listener/python/requirements.txt
requests
boto3
jsonpickle
aws-xray-sdk
cachetools
pyyaml